% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{EIA_cat_api}
\alias{EIA_cat_api}
\title{Get information on categories from the EIA API}
\usage{
EIA_cat_api(cat, key)
}
\arguments{
\item{key}{your API key as a string}

\item{category}{ID string}
}
\description{
Get information on categories from the EIA API
}
