% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervales.R
\name{eia_intervals}
\alias{eia_intervals}
\title{Create Time Interval for Batch GET Request}
\usage{
eia_intervals(start, end, tz = "UTC", length)
}
\arguments{
\item{start}{set the starting date or time of the series using "YYYY-MM-DD"
format for date and "YYYY-MM-DDTHH" format for hourly time series}

\item{end}{set the ending date or time of the series using "YYYY-MM-DD"
format for date and "YYYY-MM-DDTHH" format for hourly time series}

\item{tz}{The time zone to use to reformat the start and end arguments,
by default set as UTC. Must be one of the R time zones, see the OlsonNames()
function for available options}

\item{length}{set the interval space, based on the number of observations per
pull}
}
\value{
a vector of "POSIXct/Date objects
}
\description{
One of the limitation of the EIA API is the length limitation
(5000 observation per call), which makes the pull of a large series convoluted.
The eia_intervals function creates time intervals that can be used to batch GET
requests using the length of the call as the time/dates offset.
}
\examples{
eia_intervals(start = "2020-01-01",
              end = "2020-10-01",
              length = 30)

}
