% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{EIA_series_api}
\alias{EIA_series_api}
\title{Get EIA data series}
\usage{
EIA_series_api(ID, key)
}
\arguments{
\item{ID}{a vector of series IDs (up to 100)}

\item{key}{your API key as a string}
}
\description{
Get EIA data series
}
