% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_directory.R
\name{EIA_series_directory}
\alias{EIA_series_directory}
\title{Get EIA series IDs which conform to certain criteria}
\usage{
EIA_series_directory(...)
}
\description{
Get EIA series IDs which conform to certain criteria
}
