% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervales.R
\name{eia_to_posix}
\alias{eia_to_posix}
\title{Convert EIA API Time Format to POSIXct Object}
\usage{
eia_to_posix(time, tz = "UTC")
}
\arguments{
\item{time}{EIA time object using "YYYY-MM-DDTHH" format}

\item{tz}{The time zone format to use to set the time argument,
by default set as UTC. Must be one of the R time zones, see the OlsonNames()
function for available time zone options}
}
\description{
An helper function for converting time objects from the EIA API
format (e.g., "YYYY-MM-DDTHH) to POSIXct object
}
