% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_category_api.R
\name{getEIAcat}
\alias{getEIAcat}
\title{Get information on categories from the EIA API}
\usage{
getEIAcat(id)
}
\arguments{
\item{id}{string EIA category ID}
}
\description{
Get information on categories from the EIA API
}
