% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{getEIAcats}
\alias{getEIAcats}
\title{Get information on multiple EIA categories}
\usage{
getEIAcats(ids)
}
\arguments{
\item{ids}{a vector of category IDs. Will result in an API call for each.}
}
\description{
Get information on multiple EIA categories
}
