% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_series_api.R
\name{getEIAseries}
\alias{getEIAseries}
\title{Get an data series(es) from the API. Up to 100 can be retrieved at a time.}
\usage{
getEIAseries(ids, return = "simple")
}
\arguments{
\item{return}{flag for type of return object: simple, content, raw, or full.}

\item{id}{vector of character strings with EIA series IDs}
}
\description{
Get an data series(es) from the API. Up to 100 can be retrieved at a time.
}
