% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{update_series_dir}
\alias{update_series_dir}
\title{Expand a directory of category API requests
Will check for flag to update}
\usage{
update_series_dir(dir)
}
\arguments{
\item{dir}{a dataframe which is a directory of category IDs to update}
}
\description{
Expand a directory of category API requests
Will check for flag to update
}
