% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{getEIAcat}
\alias{getEIAcat}
\title{Get information on categories from the EIA API}
\usage{
getEIAcat(id, deep = FALSE, return = "simple")
}
\arguments{
\item{id}{string EIA category ID}

\item{deep}{indicates whether to recursively query subcategories}

\item{return}{flag for type of return object: simple, content, raw, or full.}
}
\description{
Get information on categories from the EIA API
}
