% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIA_api_reqs.R
\name{update_cat_dir}
\alias{update_cat_dir}
\title{Update a directory of category API requests
Will check whether information is old or updated prior to sending a request}
\usage{
update_cat_dir(dir)
}
\arguments{
\item{dir}{a dataframe which is a directory of category IDs to update}
}
\description{
Update a directory of category API requests
Will check whether information is old or updated prior to sending a request
}
